Split up the network configuration into different netns
The “hades networking” command will create two additional netns, one for
authenticated users (auth) and one for unauthenticated users (unauth).

The network devices corresponding to these VLANs are moved into the
appropriate netns. For each netns there is a separate keepalived
instance running (3 keepalived instances in total).

Additionally the corresponding services (DNS and DHCP) are spawned in
the corresponding netns.

Routing and firewalling is not implemented right now. This should be
fixed in later commits.