Don't set NOT NULL for foreign tables
The NOT NULL constraints on foreign tables can't be enforced locally, they have
to be enforced by the foreign database server.

The foreign server might not do this and return NULL data.

Currently PostgreSQL does not care if the NOT NULL constraint is set,
especially since we're using the foreign table through a view, which drops the
NOT NULL constraint information anyway, but this might change in the future.

We do ensure the NOT NULL constraints we need with appropriate WHERE
conditions, we do not want these WHERE conditions to be optimized out by future
versions of PostgreSQL, because of the NOT NULL constraints set on the foreign
table.