Make PostgreSQL connection params configurable
We had the HADES_POSTGRESQL_SOCKET configuration option, but it was not used in
all necessary locations. Furthermore it was actually the path of the socket
directory and not the path of the socket.

This patch renames HADES_POSTGRESQL_SOCKET to
HADES_POSTGRESQL_SOCKET_DIRECTORY, adds a HADES_POSTGRESQL_PORT option and uses
these option for the SQLAlchemy connection URI and freeRADIUS.