Make check and compute functions classmethods
Check and compute functions are functions that are assigned to the
static_check, runtime_check and default attributes of Option classes.

If we want to define e.g. a custom check for a specific Option, that is not
meaningful for other option we can now simply add a classmethod or staticmethod
to the Option, like

class MY_OPTION(Option):
    @staticmethod
    def default(config):
        …
    @staticmethod
    def static_check(config, value):
        …