Use venv
Mixing Debian-provided Python distributions and PyPI distributions can be
problematic.

As of Debian Jessie, we can't use manylinux1 wheels for example and are
restricted to the versions of our dependencies that are included in Debian.

Note: virtualenv is not used, instead this uses venv, which is included in
Python, since 3.3.