Implement a DBus deputy service for privileged operations
Some maintenance operations require a privileged user, namely
- to reload the auth DHCP hosts a signal must be sent to the
  hades-auth-dhcp.service
- to reload the RADIUS clients a signal must be sent to the
  hades-radius.service
- to reload the ipset for alternative DNS clients, ipset must be called

Instead of running the hades-agent.service and the hades-refresh.service as
root or with sufficient elevated capabilities, we run a small daemon that
provides a DBus service instead. The daemon will read the information from the
database, write new configuration files and instruct systemd to reload the
services.