Implement networking through systemd services
Network setup in hades can a bit fragile, because running the control-network
script is all-or-nothing. Additionally, the imperative nature of this scripts
makes it difficult for the user of Hades to make changes.

This patch introduces a declarative network setup based on systemd and
ifupdown. systemd itself allows the creation of network namespaces with the
PrivateNetwork= setting. This namespace is anonymous and not mounted like the
network namespaces created with ip netns. A trick described in a systemd issue
[1] can be used to make the systemd namespace available to ip netns.
The root network namespace is also made available as a named ip netns namespace
via the same trick.

The three namespaces are henceforth represented by three distinct systemd
services and other services are put into the namespace by systemd.

The network interface setup is performed programmatically through ifupdown
templates. The ifupdown templates don't interfere with the system-wide
ifupdown, which may be used by the administrator.

The hooks under /etc/network/if-*.d are however executed. These hooks are
required for ifupdown extensions, such as bridge support.

[1] https://github.com/systemd/systemd/issues/2741