Support relative config via Python's import system
A programmatically generated package named hades_config, that is located at the
directory of the config file, is added to sys.modules and the config file is
loaded as a module of this package. If the config file itself is a directory,
the __init__.py module in the directory will be used.

This allows the use of relative imports inside the config file to work with the
usual Python semantics. Because the config package is a top-level package,
files in parent or sibling directories of the config file are not possible.

The commit is part of series to address #69. In particular this change would
allow the operator to store sensitive config options in a different file with
restricted permissions.