Set mode and group during config generation
The mode is 0750. Directories will have the execute bits set, if and only if
they have the corresponding read bit is set. Files will never have any
execute, setuid, or setgid permission bits set. Sticky bits are allowed for any
file type, setgid is allowed for directories.

The mode and group can be specified on the command-line with parameters,
otherwise defaults are used.

If no group is given, the group remains unchanged.

Addresses #69