Perform template search in multiple directories
The patch enhances the config file generation mechanism to support multiple
directories, that can override others. The mechanism is analogous to systemd's
unit file lookup and the shell's PATH search.

The template search path is traversed from left to the right, so that files and
directories on the left take precedence over their counterparts with same
relative path to their respective root directory on the template search path.
Files are overridden and directories are merged.

Like with systemd units, files and directories can be masked by creating a
symbolic link to /dev/null.

The default template search path is set to:
/etc/hades/templates:/usr/share/hades/templates