Fix rendering of `hades:option` directives
The `add_target_and_index` code has (at least in current sphinx
versions) never been executed until now, because we need to implement
`handle_signature`. Otherwise, creation of an index node by calling
`add_target_and_index()` is not attempted.

In the course of this, several minor things have been fixed.

As a result, the option nodes are now properly rendered with a
permalink symbol instead of the plain fallback.